requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
SQLAlchemy>=2.0,<3.0
alembic>=1.13,<2.0
psycopg[binary]>=3.1,<4.0
pydantic-settings>=2.2,<3.0
python-dotenv>=1.0,<2.0
python-multipart>=0.0.9,<1.0
pdfplumber>=0.11,<1.0
pypdfium2>=4.30,<5.0
pytesseract>=0.3.10,<1.0
Pillow>=10.0,<12.0
celery[redis]>=5.4,<6.0
openpyxl>=3.1,<4.0
